Fix bug when checking `n` for `NA`
The current check errors when `n` is a vector, this fix just adds `any()` to the check to see if any values for `n` are `NA` in the vector.

It also cleans up the `stop()` function. The use of `paste0()` is redundant for the message returned to users, and added `call. = FALSE` for a cleaner stop message when a `stop()` is encountered.

The package passes checks with 1 Warning (deprecated fn `default.stringsAsFactors`) and one Note