Use Rd macros
Use \CRANpkg{} in place of \strong{} for the _epiphy_ name in documentation.

Use \email{} in place of `<>` as per https://cran.r-project.org/doc/manuals/R-exts.html,
>\author{…}
Information about the author(s) of the Rd file. Use \email{} without extra delimiters (such as ‘( )’ or ‘< >’) to specify email addresses, or \url{} or \href{}{} for web pointers.